brevo/requirements.txt
# Dépendances pour la collecte des événements marketing Brevo

# HTTP requests
requests>=2.31.0

# YAML configuration
PyYAML>=6.0.1

# Google Cloud BigQuery
google-cloud-bigquery>=3.11.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
